Fix archive bug on reinsert same foreign key.
Previously it would attempt to start over with revision 0 if the same
primary key was used in the parent table. This would run into key
problems on the archive table causing inserts and updates to fail. This
update causes re-insert of the same primary key in a table to use the
last revision count + 1.